# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-PICSA Climate API",
    "description": "\n### E-PICSA Climate API Documentation\n\n* **v2 Endpoints**: High-performance endpoints querying PostgreSQL directly. [View v2 Docs Only](/v2/docs)\n* **v1 Endpoints (Legacy)**: Legacy endpoints using R (`epicsawrap`) and Cloud Storage. [View v1 Docs Only](/v1/docs)\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/v1/select_query/": {
      "post": {
        "tags": [
          "v1 (Legacy): Database"
        ],
        "summary": "Run Select Query",
        "operationId": "run_select_query_v1_select_query__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/select_query/": {
      "post": {
        "tags": [
          "v2: Database"
        ],
        "summary": "Run Select Query",
        "operationId": "run_select_query_v2_select_query__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SelectQueryRequest": {
        "properties": {
          "table_name": {
            "type": "string",
            "enum": [
              "crop",
              "definition",
              "station",
              "summary",
              "summary_station_metadata"
            ],
            "title": "Table Name"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns",
            "description": "Optional list of allowed columns. Empty means all allowed columns."
          },
          "station_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Station Id"
          },
          "order_by": {
            "type": "string",
            "title": "Order By",
            "description": "Optional sort column. Must be allowed for the selected table."
          },
          "order_direction": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Order Direction",
            "default": "desc"
          },
          "max_rows": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Rows",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "table_name",
          "station_id"
        ],
        "title": "SelectQueryRequest"
      },
      "SelectQueryResponse": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "rows": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "columns",
          "row_count",
          "rows"
        ],
        "title": "SelectQueryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}